feat: implement modular API routes for projects and legacy operations
